+++ qmetaobject/tests/tests.rs
@@ -979,3 +979,28 @@ fn test_qvariant_qimage_qpixmap() {
     img3.set_pixel_color(8, 8, QColor::from_name("black"));
     assert!(img2 != img3);
 }
+
+#[test]
+fn test_application_name() {
+    use qmetaobject::qtcore::core_application::*;
+
+    let app_name    = QString::from("qmetaobject-rs testing app");
+    let app_version = QString::from("0.1");
+    let org_domain  = QString::from("woboq.com");
+    let org_name    = QString::from("Woboq");
+
+    QCoreApplication::set_application_name(app_name.clone());
+    QCoreApplication::set_application_version(app_version.clone());
+    QCoreApplication::set_organization_domain(org_domain.clone());
+    QCoreApplication::set_organization_name(org_name.clone());
+
+    let app_name_    = QCoreApplication::application_name();
+    let app_version_ = QCoreApplication::application_version();
+    let org_domain_  = QCoreApplication::organization_domain();
+    let org_name_    = QCoreApplication::organization_name();
+
+    assert_eq!(app_name,    app_name_);
+    assert_eq!(app_version, app_version_);
+    assert_eq!(org_domain,  org_domain_);
+    assert_eq!(org_name,    org_name_);
+}